234: copyEPackages in the derived state helper
Task-Url: http://github.com/LorenzoBettini/edelta/issues/234
